+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 #### date
 
 - _hexagonal binning_: first version
+- _box plot_: added colors, labels and styles support
+- _box plot_: it is possible to set box width
+- _box plot_: `values` dimension renamed to `Y Axis`.
+- _box plot_: added "Start from 0" visual option.
+- _stream graph_: first version
 
 #### 10 feb 2021